feat(wayfinder): add tapResponseStream helper, add events
We need to create a more generic mechanism for tapping Readables and ReadableStreams vs. relying on .clone() - which is only available on the Repsonse type. These methods should properly create tapped streams, one needed for verification and the other needed to give to the client. Due to back pressure contraints, we cannot just consume the verification stream independenly of the client. This means verification starts once the client starts consuming the response, and if it fails, the client stream is then cancelled/stopped before the receive the final bytes. At least that is the intent. More tests to do with various http clients, but progress nonetheless.