feat(ants): add support for fetching ant state from hyperbeam nodes
This is an early implementation that needs some tuning, but the intent is with newer ANTs publishing patch messages of their state on changes, we can now start to lean in to using hyperbeam for cached state and lower the need for CUs. We may want to modify how we are verifying HyperBeam compatability vs. the HEAD check against a known hyperbeam node.'